code refactoring, "isLCYClaimed" is a field of the card item now.
