+++ src/store/state.js
@@ -11,7 +11,6 @@ export default () => ({
   locale: '',
   me: null,
   signInError: null,
-  isClaimedLCY: false,
   items: {},
   ads: {},
   /* Example
