+++ src/views/Identicon.js
